+++ benches/history_buf.rs
@@ -3,31 +3,25 @@ use std::hint::black_box;
 use divan::counter::BytesCount;
 use heapless::HistoryBuf;
 
-const NEEDLE: &[u8] = b"needle451";
-const WINDOW_SIZE: usize = NEEDLE.len();
+const WINDOW_SIZE: usize = 9;
 
-fn write_search_window(data: Vec<u8>) {
+fn write_search_window(data: &[u8]) {
     let mut search_window: HistoryBuf<u8, WINDOW_SIZE> = HistoryBuf::new();
 
-    for byte in data {
+    for byte in data.iter().copied() {
         search_window.write(byte);
     }
+    black_box(&search_window);
 }
 
 #[divan::bench]
 fn history_buf_write(bencher: divan::Bencher) {
     let total_bytes = 8 * 1024 * 1024;
-    let bytes_before_needle = total_bytes - NEEDLE.len();
 
     bencher
         .counter(BytesCount::new(total_bytes))
-        .with_inputs(|| {
-            let mut data = Vec::with_capacity(total_bytes);
-            data.resize(bytes_before_needle, 0);
-            data.extend(NEEDLE);
-            data
-        })
-        .bench_values(|data| write_search_window(black_box(data)));
+        .with_inputs(|| vec![0u8; total_bytes])
+        .bench_refs(|data| write_search_window(black_box(data)));
 }
 
 fn main() {